住民税を追加。従来の控除関係の実装を整理
