+++ openfisca_japan/parameters/住民税/均等割.yaml
@@ -1,0 +1,7 @@
+description: 住民税における均等割
+metadata:
+  reference: https://www.soumu.go.jp/main_sosiki/jichi_zeisei/czaisei/czaisei_seido/150790_06.html
+  unit: currency-JPY
+values:
+  2000-01-01:
+    value: 4000
