+++ openfisca_japan/tests/住民税/住民税.yaml
@@ -1,0 +1,43 @@
+- name: 住民税（通常ケース）
+  absolute_error_margin: 1
+  period: 2024-06-01
+  input:
+    世帯:
+      親一覧:
+        - 親1
+    世帯員:
+      親1:
+        住民税所得割: 100000
+        住民税均等割: 5000
+  output:
+    世帯員:
+      親1:
+        住民税:
+          '2024-06-01': 105000 # 100000 + 5000
+
+- name: 住民税（複数世帯）
+  absolute_error_margin: 1
+  period: 2024-06-01
+  input:
+    世帯一覧:
+      世帯1:
+        親一覧:
+          - 親1
+      世帯2:
+        親一覧:
+          - 親2
+    世帯員:
+      親1:
+        住民税所得割: 50000
+        住民税均等割: 5000
+      親2:
+        住民税所得割: 150000
+        住民税均等割: 5000
+  output:
+    世帯員:
+      親1:
+        住民税:
+          '2024-06-01': 55000
+      親2:
+        住民税:
+          '2024-06-01': 155000
