+++ openfisca_japan/tests/住民税/住民税均等割.yaml
@@ -1,0 +1,37 @@
+- name: 住民税均等割（全世帯一律）
+  absolute_error_margin: 1
+  period: 2024-06-01
+  input:
+    世帯:
+      親一覧:
+        - 親1
+    世帯員:
+      親1: {}
+  output:
+    世帯員:
+      親1:
+        住民税均等割:
+          '2024-06-01': 5000 # 4000 (均等割) + 1000 (森林環境税)
+
+- name: 住民税均等割（複数世帯）
+  absolute_error_margin: 1
+  period: 2024-06-01
+  input:
+    世帯一覧:
+      世帯1:
+        親一覧:
+          - 親1
+      世帯2:
+        親一覧:
+          - 親2
+    世帯員:
+      親1: {}
+      親2: {}
+  output:
+    世帯員:
+      親1:
+        住民税均等割:
+          '2024-06-01': 5000
+      親2:
+        住民税均等割:
+          '2024-06-01': 5000
