+++ openfisca_japan/tests/住民税/住民税所得割.yaml
@@ -1,0 +1,64 @@
+- name: 住民税所得割（通常ケース）
+  absolute_error_margin: 1
+  period: 2024-06-01
+  input:
+    世帯:
+      親一覧:
+        - 親1
+      調整控除: 2500
+    世帯員:
+      親1:
+        合計所得金額: 3000000
+        住民税所得控除: 1750000
+  output:
+    世帯員:
+      親1:
+        住民税所得割:
+          '2024-06-01': 122500 # (3000000 - 1750000) * 0.1 - 2500 = 125000 - 2500
+
+- name: 住民税所得割（非課税）
+  absolute_error_margin: 1
+  period: 2024-06-01
+  input:
+    世帯:
+      親一覧:
+        - 親1
+      調整控除: 2500
+    世帯員:
+      親1:
+        合計所得金額: 1000000
+        住民税所得控除: 1120000
+  output:
+    世帯員:
+      親1:
+        住民税所得割:
+          '2024-06-01': 0 # Negative result clipped to 0
+
+- name: 住民税所得割（複数世帯）
+  absolute_error_margin: 1
+  period: 2024-06-01
+  input:
+    世帯一覧:
+      世帯1:
+        親一覧:
+          - 親1
+        調整控除: 2000
+      世帯2:
+        親一覧:
+          - 親2
+        調整控除: 5000
+    世帯員:
+      親1:
+        合計所得金額: 2000000
+        住民税所得控除: 1000000
+      親2:
+        合計所得金額: 4000000
+        住民税所得控除: 2000000
+  output:
+    世帯員:
+      親1:
+        住民税所得割:
+          '2024-06-01': 98000 # (2000000 - 1000000) * 0.1 - 2000 = 100000 - 2000
+      親2:
+        住民税所得割:
+          '2024-06-01': 195000 # (4000000 - 2000000) * 0.1 - 5000 = 200000 - 5000
