+++ openfisca_japan/tests/住民税/住民税所得控除.yaml
@@ -1,0 +1,64 @@
+- name: 住民税所得控除（全項目合算）
+  absolute_error_margin: 1
+  period: 2024-06-01
+  input:
+    世帯:
+      親一覧:
+        - 親1
+    世帯員:
+      親1:
+        社会保険料控除: 100000
+        住民税障害者控除: 260000
+        住民税寡婦控除: 0
+        住民税ひとり親控除: 300000
+        住民税勤労学生控除: 0
+        住民税扶養控除: 330000
+        住民税配偶者控除: 330000
+        住民税配偶者特別控除: 0
+        住民税基礎控除: 430000
+  output:
+    世帯員:
+      親1:
+        住民税所得控除:
+          '2024-06-01': 1750000 # 100000 + 260000 + 300000 + 330000 + 330000 + 430000
+
+- name: 住民税所得控除（複数世帯）
+  absolute_error_margin: 1
+  period: 2024-06-01
+  input:
+    世帯一覧:
+      世帯1:
+        親一覧:
+          - 親1
+      世帯2:
+        親一覧:
+          - 親2
+    世帯員:
+      親1:
+        社会保険料控除: 50000
+        住民税障害者控除: 0
+        住民税寡婦控除: 0
+        住民税ひとり親控除: 0
+        住民税勤労学生控除: 0
+        住民税扶養控除: 0
+        住民税配偶者控除: 0
+        住民税配偶者特別控除: 0
+        住民税基礎控除: 430000
+      親2:
+        社会保険料控除: 100000
+        住民税障害者控除: 0
+        住民税寡婦控除: 260000
+        住民税ひとり親控除: 0
+        住民税勤労学生控除: 0
+        住民税扶養控除: 330000
+        住民税配偶者控除: 0
+        住民税配偶者特別控除: 0
+        住民税基礎控除: 430000
+  output:
+    世帯員:
+      親1:
+        住民税所得控除:
+          '2024-06-01': 480000
+      親2:
+        住民税所得控除:
+          '2024-06-01': 1120000
