+++ openfisca_japan/tests/住民税/各種控除.yaml
@@ -817,12 +817,13 @@
     世帯員:
       親1:
         誕生年月日: '1993-05-01'
-        所得:
+        合計所得金額:
           '2023-06-01': 24000000
   output:
-    世帯:
-      住民税基礎控除:
-        '2023-06-01': 430000
+    世帯員:
+      親1:
+        住民税基礎控除:
+          '2023-06-01': 430000
 
 - name: 住民税における基礎控除2
   period: 2023-06-01
@@ -833,12 +834,13 @@
     世帯員:
       親1:
         誕生年月日: '1993-05-01'
-        所得:
+        合計所得金額:
           '2023-06-01': 24000002 # NOTE: float32の丸め誤差で1円差は無視されてしまう
   output:
-    世帯:
-      住民税基礎控除:
-        '2023-06-01': 290000
+    世帯員:
+      親1:
+        住民税基礎控除:
+          '2023-06-01': 290000
 
 - name: 住民税における基礎控除3
   period: 2023-06-01
@@ -849,12 +851,13 @@
     世帯員:
       親1:
         誕生年月日: '1993-05-01'
-        所得:
+        合計所得金額:
           '2023-06-01': 24500000
   output:
-    世帯:
-      住民税基礎控除:
-        '2023-06-01': 290000
+    世帯員:
+      親1:
+        住民税基礎控除:
+          '2023-06-01': 290000
 
 - name: 住民税における基礎控除4
   period: 2023-06-01
@@ -865,12 +868,13 @@
     世帯員:
       親1:
         誕生年月日: '1993-05-01'
-        所得:
+        合計所得金額:
           '2023-06-01': 24500002 # NOTE: float32の丸め誤差で1円差は無視されてしまう
   output:
-    世帯:
-      住民税基礎控除:
-        '2023-06-01': 150000
+    世帯員:
+      親1:
+        住民税基礎控除:
+          '2023-06-01': 150000
 
 - name: 住民税における基礎控除5
   period: 2023-06-01
@@ -881,12 +885,13 @@
     世帯員:
       親1:
         誕生年月日: '1993-05-01'
-        所得:
+        合計所得金額:
           '2023-06-01': 25000000
   output:
-    世帯:
-      住民税基礎控除:
-        '2023-06-01': 150000
+    世帯員:
+      親1:
+        住民税基礎控除:
+          '2023-06-01': 150000
 
 - name: 住民税における基礎控除6
   period: 2023-06-01
@@ -897,12 +902,13 @@
     世帯員:
       親1:
         誕生年月日: '1993-05-01'
-        所得:
+        合計所得金額:
           '2023-06-01': 25000002 # NOTE: float32の丸め誤差で1円差は無視されてしまう
   output:
-    世帯:
-      住民税基礎控除:
-        '2023-06-01': 0
+    世帯員:
+      親1:
+        住民税基礎控除:
+          '2023-06-01': 0
 
 - name: 住民税における基礎控除（複数世帯）
   period: 2023-06-01
@@ -917,18 +923,18 @@
     世帯員:
       親1:
         誕生年月日: '1993-05-01'
-        所得:
+        合計所得金額:
           '2023-06-01': 24000000
       親2:
         誕生年月日: '1993-05-01'
-        所得:
+        合計所得金額:
           '2023-06-01': 24000002 # NOTE: float32の丸め誤差で1円差は無視されてしまう
   output:
-    世帯一覧:
-      世帯1:
+    世帯員:
+      親1:
         住民税基礎控除:
           '2023-06-01': 430000
-      世帯2:
+      親2:
         住民税基礎控除:
           '2023-06-01': 290000
 
@@ -947,21 +953,21 @@
     世帯員:
       親1:
         誕生年月日: '1993-05-01'
-        所得:
+        合計所得金額:
           '2023-06-01': 24000000
       子1:
         誕生年月日: '2013-05-01'
-        所得:
+        合計所得金額:
           '2023-06-01': 0
       親2:
         誕生年月日: '1993-05-01'
-        所得:
+        合計所得金額:
           '2023-06-01': 24000002 # NOTE: float32の丸め誤差で1円差は無視されてしまう
   output:
-    世帯一覧:
-      世帯1:
+    世帯員:
+      親1:
         住民税基礎控除:
           '2023-06-01': 430000
-      世帯2:
+      親2:
         住民税基礎控除:
           '2023-06-01': 290000
