+++ openfisca_japan/tests/住民税/調整控除.yaml
@@ -286,10 +286,12 @@
     世帯:
       親一覧:
         - 自分
+      人的控除額の差:
+        '2023-06-01': 20000
+      控除後住民税世帯高所得:
+        '2023-06-01': 10000
     世帯員:
       自分:
-        所得:
-          '2023-06-01': 440000 # 控除後世帯高所得は 440000 - 430000(基礎控除) = 10000
         誕生年月日: '1983-01-01'
   output:
     世帯:
@@ -302,10 +304,12 @@
     世帯:
       親一覧:
         - 自分
+      人的控除額の差:
+        '2023-06-01': 50000
+      控除後住民税世帯高所得:
+        '2023-06-01': 2010000
     世帯員:
       自分:
-        所得:
-          '2023-06-01': 2440000 # 控除後世帯高所得は 2440000 - 430000(基礎控除) = 2010000
         誕生年月日: '1983-01-01'
   output:
     世帯:
