+++ openfisca_japan/tests/所得/所得税.yaml
@@ -1,0 +1,337 @@
+- name: 基準所得税額（課税所得1,950,000未満_上限1,949,999）
+  absolute_error_margin: 1
+  period: 2025-01-01
+  input:
+    世帯:
+      親一覧:
+        - 親1
+    世帯員:
+      親1:
+        合計所得金額: 1949999
+        所得控除: 0
+  output:
+    世帯員:
+      親1:
+        基準所得税額:
+          '2025-01-01': 97499 # 1,949,999 * 0.05
+
+
+- name: 基準所得税額（課税所得1,950,000以上 3,300,000未満_下限1,950,000）
+  absolute_error_margin: 1
+  period: 2025-01-01
+  input:
+    世帯:
+      親一覧:
+        - 親1
+    世帯員:
+      親1:
+        合計所得金額: 1950000
+        所得控除: 0
+  output:
+    世帯員:
+      親1:
+        基準所得税額:
+          '2025-01-01': 97500 # 1,950,000 * 0.1 - 97,500
+
+- name: 基準所得税額（課税所得1,950,000以上 3,300,000未満_上限3,299,999）
+  absolute_error_margin: 1
+  period: 2025-01-01
+  input:
+    世帯:
+      親一覧:
+        - 親1
+    世帯員:
+      親1:
+        合計所得金額: 3299999
+        所得控除: 0
+  output:
+    世帯員:
+      親1:
+        基準所得税額:
+          '2025-01-01': 232499 # 3,299,999 * 0.1 - 97,500
+
+- name: 基準所得税額（課税所得3,300,000以上 6,950,000未満_下限3,300,000）
+  absolute_error_margin: 1
+  period: 2025-01-01
+  input:
+    世帯:
+      親一覧:
+        - 親1
+    世帯員:
+      親1:
+        合計所得金額: 3300000
+        所得控除: 0
+  output:
+    世帯員:
+      親1:
+        基準所得税額:
+          '2025-01-01': 232500 # 3,300,000 * 0.2 - 427,500
+
+- name: 基準所得税額（課税所得3,300,000以上 6,950,000未満_上限6,949,999）
+  absolute_error_margin: 1
+  period: 2025-01-01
+  input:
+    世帯:
+      親一覧:
+        - 親1
+    世帯員:
+      親1:
+        合計所得金額: 6949999
+        所得控除: 0
+  output:
+    世帯員:
+      親1:
+        基準所得税額:
+          '2025-01-01': 962499 # 6,949,999 * 0.2 - 427,500
+
+- name: 基準所得税額（課税所得6,950,000以上 9,000,000未満_下限6,950,000）
+  absolute_error_margin: 1
+  period: 2025-01-01
+  input:
+    世帯:
+      親一覧:
+        - 親1
+    世帯員:
+      親1:
+        合計所得金額: 6950000
+        所得控除: 0
+  output:
+    世帯員:
+      親1:
+        基準所得税額:
+          '2025-01-01': 962500 # 6,950,000 * 0.23 - 636,000
+
+- name: 基準所得税額（課税所得6,950,000以上 9,000,000未満_上限8,999,999）
+  absolute_error_margin: 1
+  period: 2025-01-01
+  input:
+    世帯:
+      親一覧:
+        - 親1
+    世帯員:
+      親1:
+        合計所得金額: 8999999
+        所得控除: 0
+  output:
+    世帯員:
+      親1:
+        基準所得税額:
+          '2025-01-01': 1433999 # 8,999,999 * 0.23 - 636,000
+
+- name: 基準所得税額（課税所得9,000,000以上 18,000,000未満_下限9,000,000）
+  absolute_error_margin: 1
+  period: 2025-01-01
+  input:
+    世帯:
+      親一覧:
+        - 親1
+    世帯員:
+      親1:
+        合計所得金額: 9000000
+        所得控除: 0
+  output:
+    世帯員:
+      親1:
+        基準所得税額:
+          '2025-01-01': 1434000 # 9,000,000 * 0.33 - 1,536,000
+
+- name: 基準所得税額（課税所得9,000,000以上 18,000,000未満_上限17,999,999）
+  absolute_error_margin: 1
+  period: 2025-01-01
+  input:
+    世帯:
+      親一覧:
+        - 親1
+    世帯員:
+      親1:
+        合計所得金額: 17999999
+        所得控除: 0
+  output:
+    世帯員:
+      親1:
+        基準所得税額:
+          '2025-01-01': 4403999 # 17,999,999 * 0.33 - 1,536,000
+
+- name: 基準所得税額（課税所得18,000,000以上 40,000,000未満_下限18,000,000）
+  absolute_error_margin: 1
+  period: 2025-01-01
+  input:
+    世帯:
+      親一覧:
+        - 親1
+    世帯員:
+      親1:
+        合計所得金額: 18000000
+        所得控除: 0
+  output:
+    世帯員:
+      親1:
+        基準所得税額:
+          '2025-01-01': 4404000 # 18,000,000 * 0.40 - 2,796,000
+
+- name: 基準所得税額（課税所得18,000,000以上 40,000,000未満_上限39,999,999）
+  absolute_error_margin: 1
+  period: 2025-01-01
+  input:
+    世帯:
+      親一覧:
+        - 親1
+    世帯員:
+      親1:
+        合計所得金額: 39999999
+        所得控除: 0
+  output:
+    世帯員:
+      親1:
+        基準所得税額:
+          '2025-01-01': 13203999 # 39,999,999 * 0.40 - 2,796,000
+
+- name: 基準所得税額（課税所得40,000,000以上）
+  absolute_error_margin: 1
+  period: 2025-01-01
+  input:
+    世帯:
+      親一覧:
+        - 親1
+    世帯員:
+      親1:
+        合計所得金額: 40000000
+        所得控除: 0
+  output:
+    世帯員:
+      親1:
+        基準所得税額:
+          '2025-01-01': 13204000 # 40,000,000 * 0.45 - 4,796,000
+
+- name: 基準所得税額（複数世帯）
+  absolute_error_margin: 1
+  period: 2025-01-01
+  input:
+    世帯一覧:
+      世帯1:
+        親一覧:
+          - 親1
+      世帯2:
+        親一覧:
+          - 親2
+    世帯員:
+      親1:
+        合計所得金額: 1950000
+        所得控除: 0
+      親2:
+        合計所得金額: 3300000
+        所得控除: 0
+  output:
+    世帯員:
+      親1:
+        基準所得税額:
+          '2025-01-01': 97500
+      親2:
+        基準所得税額:
+          '2025-01-01': 232500
+
+- name: 復興特別所得税（通常ケース）
+  absolute_error_margin: 1
+  period: 2025-01-01
+  input:
+    世帯:
+      親一覧:
+        - 親1
+    世帯員:
+      親1:
+        基準所得税額: 100000
+  output:
+    世帯員:
+      親1:
+        復興特別所得税:
+          '2025-01-01': 2100 # 100,000 * 0.021
+
+- name: 復興特別所得税（複数世帯）
+  absolute_error_margin: 1
+  period: 2025-01-01
+  input:
+    世帯一覧:
+      世帯1:
+        親一覧:
+          - 親1
+      世帯2:
+        親一覧:
+          - 親2
+    世帯員:
+      親1:
+        基準所得税額: 100000
+      親2:
+        基準所得税額: 200000
+  output:
+    世帯員:
+      親1:
+        復興特別所得税:
+          '2025-01-01': 2100
+      親2:
+        復興特別所得税:
+          '2025-01-01': 4200
+
+- name: 所得税（超過所得税適用なし）
+  absolute_error_margin: 1
+  period: 2025-01-01
+  input:
+    世帯:
+      親一覧:
+        - 親1
+    世帯員:
+      親1:
+        基準所得税額: 100000
+        復興特別所得税: 2100
+        合計所得金額: 10000000
+  output:
+    世帯員:
+      親1:
+        所得税:
+          '2025-01-01': 102100 # 100,000 + 2,100
+
+- name: 所得税（超過所得税適用あり）
+  absolute_error_margin: 1
+  period: 2025-01-01
+  input:
+    世帯:
+      親一覧:
+        - 親1
+    世帯員:
+      親1:
+        基準所得税額: 10000000
+        復興特別所得税: 210000
+        合計所得金額: 380000000
+  output:
+    世帯員:
+      親1:
+        所得税:
+          '2025-01-01': 11250000 # (380,000,000 - 330,000,000) * 0.225 = 11,250,000 > 10,210,000
+
+- name: 所得税（複数世帯）
+  absolute_error_margin: 1
+  period: 2025-01-01
+  input:
+    世帯一覧:
+      世帯1:
+        親一覧:
+          - 親1
+      世帯2:
+        親一覧:
+          - 親2
+    世帯員:
+      親1:
+        基準所得税額: 100000
+        復興特別所得税: 2100
+        合計所得金額: 10000000
+      親2:
+        基準所得税額: 10000000
+        復興特別所得税: 210000
+        合計所得金額: 380000000
+  output:
+    世帯員:
+      親1:
+        所得税:
+          '2025-01-01': 102100
+      親2:
+        所得税:
+          '2025-01-01': 11250000
