+++ openfisca_japan/tests/福祉/育児/障害児福祉手当.yaml
@@ -1,5 +1,5 @@
 ### TODO: 自動生成した「openfisca_japan/tests/generated/子育て支援制度テスト.yml」に単数世帯のテストが記載されている
-### 控除後世帯高所得の扱い含めてマージを検討
+### 障害児福祉手当の控除後世帯高所得の扱い含めてマージを検討
 
 - name: 複数世帯テスト
   period: 2023-07-01
@@ -10,14 +10,14 @@
           - 自分
         子一覧:
           - 子1_1
-        控除後世帯高所得: 0
+        障害児福祉手当の控除後世帯高所得: 0
       世帯2:
         親一覧:
           - 自分2
         子一覧:
           - 子2_1
           - 子2_2
-        控除後世帯高所得: 6748999
+        障害児福祉手当の控除後世帯高所得: 6748999
     世帯員:
       自分:
         年齢: 30
