+++ openfisca_japan/variables/福祉/育児/障害児福祉手当.py
@@ -50,7 +50,7 @@ class 障害児福祉手当(Variable):
         受給者の所得条件 = 対象障害者の所得 < 受給者の所得制限限度額
 
         # 世帯高所得は保護者によるものだと想定する
-        世帯高所得 = 対象世帯("控除後世帯高所得", 対象期間)
+        世帯高所得 = 対象世帯("障害児福祉手当の控除後世帯高所得", 対象期間)
         扶養義務者の所得制限限度額 = 対象世帯("障害児福祉手当_扶養義務者の所得制限限度額", 対象期間)
         扶養義務者の所得条件 = 世帯高所得 < 扶養義務者の所得制限限度額
